+++ user-operator/src/main/java/io/strimzi/operator/user/Main.java
@@ -15,7 +15,6 @@ import io.strimzi.certs.OpenSslCertManager;
 import io.strimzi.operator.common.operator.resource.CrdOperator;
 import io.strimzi.operator.common.operator.resource.SecretOperator;
 import io.strimzi.operator.user.operator.KafkaUserOperator;
-import io.strimzi.operator.user.operator.KafkaUserQuotas;
 import io.strimzi.operator.user.operator.KafkaUserQuotasOperator;
 import io.strimzi.operator.user.operator.ScramShaCredentials;
 import io.strimzi.operator.user.operator.ScramShaCredentialsOperator;
@@ -65,8 +64,7 @@ public class Main {
         SimpleAclOperator aclOperations = new SimpleAclOperator(vertx, authorizer);
         ScramShaCredentials scramShaCredentials = new ScramShaCredentials(config.getZookeperConnect(), (int) config.getZookeeperSessionTimeoutMs());
         ScramShaCredentialsOperator scramShaCredentialsOperator = new ScramShaCredentialsOperator(vertx, scramShaCredentials);
-        KafkaUserQuotas kafkaUserQuotas = new KafkaUserQuotas(config.getZookeperConnect(), (int) config.getZookeeperSessionTimeoutMs());
-        KafkaUserQuotasOperator quotasOperator = new KafkaUserQuotasOperator(vertx, kafkaUserQuotas);
+        KafkaUserQuotasOperator quotasOperator = new KafkaUserQuotasOperator(vertx, config.getZookeperConnect(), (int) config.getZookeeperSessionTimeoutMs());
 
         KafkaUserOperator kafkaUserOperations = new KafkaUserOperator(vertx,
                 certManager, crdOperations,
