+++ user-operator/src/test/java/io/strimzi/operator/user/operator/KafkaUserQuotasIT.java
@@ -4,7 +4,9 @@
  */
 package io.strimzi.operator.user.operator;
 
+import io.strimzi.api.kafka.model.KafkaUserQuotas;
 import io.strimzi.test.EmbeddedZooKeeper;
+import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
@@ -24,9 +26,9 @@ public class KafkaUserQuotasIT {
 
     private static EmbeddedZooKeeper zkServer;
 
-    private KafkaUserQuotas kuq;
+    private KafkaUserQuotasOperator kuq;
 
-    private JsonObject quotasJson;
+    private KafkaUserQuotas defaultQuotas;
 
 
     @BeforeAll
@@ -41,21 +43,21 @@ public class KafkaUserQuotasIT {
 
     @BeforeEach
     public void createKUQ() {
-        quotasJson = new JsonObject();
-        quotasJson.put("consumer_byte_rate", "1000");
-        quotasJson.put("producer_byte_rate", "2000");
-        kuq = new KafkaUserQuotas(zkServer.getZkConnectString(), 6_000);
+        defaultQuotas = new KafkaUserQuotas();
+        defaultQuotas.setConsumerByteRate(1000);
+        defaultQuotas.setProducerByteRate(2000);
+        kuq = new KafkaUserQuotasOperator(Vertx.vertx(), zkServer.getZkConnectString(), 6_000);
     }
 
     @Test
     public void normalCreate() {
-        JsonObject quotasJson = new JsonObject();
         assertThat(kuq.exists("normalCreate"), is(false));
         assertThat(kuq.getQuotas("normalCreate"), is(nullValue()));
         assertThat(kuq.isPathExist("/config/users/normalCreate"), is(false));
-        quotasJson.put("consumer_byte_rate", "1000");
-        quotasJson.put("producer_byte_rate", "2000");
-        kuq.createOrUpdate("normalCreate", quotasJson);
+        KafkaUserQuotas newQuotas = new KafkaUserQuotas();
+        newQuotas.setConsumerByteRate(1000);
+        newQuotas.setProducerByteRate(2000);
+        kuq.createOrUpdate("normalCreate", newQuotas);
         assertThat(kuq.exists("normalCreate"), is(true));
         assertThat(kuq.getQuotas("normalCreate").getJsonObject("config").getString("consumer_byte_rate"), is("1000"));
         assertThat(kuq.getQuotas("normalCreate").getJsonObject("config").getString("producer_byte_rate"), is("2000"));
@@ -67,8 +69,8 @@ public class KafkaUserQuotasIT {
         assertThat(kuq.isPathExist("/config/users/doublelCreate"), is(false));
         assertThat(kuq.exists("doubleCreate"), is(false));
         assertThat(kuq.getQuotas("doubleCreate"), is(nullValue()));
-        kuq.createOrUpdate("doubleCreate", quotasJson);
-        kuq.createOrUpdate("doubleCreate", quotasJson);
+        kuq.createOrUpdate("doubleCreate", defaultQuotas);
+        kuq.createOrUpdate("doubleCreate", defaultQuotas);
         assertThat(kuq.exists("doubleCreate"), is(true));
         assertThat(kuq.getQuotas("doubleCreate").getJsonObject("config").getString("consumer_byte_rate"), is("1000"));
         assertThat(kuq.getQuotas("doubleCreate").getJsonObject("config").getString("producer_byte_rate"), is("2000"));
@@ -77,7 +79,7 @@ public class KafkaUserQuotasIT {
 
     @Test
     public void normalDelete() {
-        kuq.createOrUpdate("normalDelete", quotasJson);
+        kuq.createOrUpdate("normalDelete", defaultQuotas);
         assertThat(kuq.isPathExist("/config/users/normalDelete"), is(true));
         assertThat(kuq.exists("normalDelete"), is(true));
         kuq.delete("normalDelete");
@@ -86,7 +88,7 @@ public class KafkaUserQuotasIT {
 
     @Test
     public void doubleDelete() {
-        kuq.createOrUpdate("doubleDelete", quotasJson);
+        kuq.createOrUpdate("doubleDelete", defaultQuotas);
         assertThat(kuq.isPathExist("/config/users/doubleDelete"), is(true));
         assertThat(kuq.exists("doubleDelete"), is(true));
         kuq.delete("doubleDelete");
@@ -95,16 +97,15 @@ public class KafkaUserQuotasIT {
     }
 
     @Test
-    public void changePassword() {
-        kuq.createOrUpdate("changePassword", quotasJson);
-        quotasJson.put("producer_byte_rate", "8000");
-        kuq.createOrUpdate("changePassword", quotasJson);
-        quotasJson.remove("producer_byte_rate");
+    public void changeProducerByteRate() {
+        kuq.createOrUpdate("changeProducerByteRate", defaultQuotas);
+        defaultQuotas.setProducerByteRate(8000);
+        kuq.createOrUpdate("changeProducerByteRate", defaultQuotas);
     }
 
     @Test
     public void userExists() {
-        kuq.createOrUpdate("userExists", quotasJson);
+        kuq.createOrUpdate("userExists", defaultQuotas);
         assertThat(kuq.exists("userExists"), is(true));
     }
 
@@ -139,7 +140,7 @@ public class KafkaUserQuotasIT {
 
     @Test
     public void testDeletion()  {
-        JsonObject original = new JsonObject().put("version", 1).put("config", quotasJson);
+        JsonObject original = new JsonObject().put("version", 1).put("config", kuq.quotasToJson(defaultQuotas));
         JsonObject updated = new JsonObject(new String(kuq.removeQuotasFromJsonUser(original.encode().getBytes(Charset.defaultCharset())), Charset.defaultCharset()));
         assertThat(updated.getJsonObject("config").getString("consumer_byte_rate"), is(nullValue()));
 
@@ -162,19 +163,25 @@ public class KafkaUserQuotasIT {
     @Test
     public void testUpdate()  {
         JsonObject original = new JsonObject().put("version", 1).put("config", new JsonObject().put("consumer_byte_rate", "1000"));
-        JsonObject updated = new JsonObject(new String(kuq.updateUserJson(original.encode().getBytes(Charset.defaultCharset()), new JsonObject().put("consumer_byte_rate", "2000")), Charset.defaultCharset()));
+        KafkaUserQuotas quotas = new KafkaUserQuotas();
+        quotas.setConsumerByteRate(2000);
+        JsonObject updated = new JsonObject(new String(kuq.updateUserJson(original.encode().getBytes(Charset.defaultCharset()), quotas), Charset.defaultCharset()));
         assertThat(updated.getJsonObject("config").getString("consumer_byte_rate"), is(notNullValue()));
         assertThat(updated.getJsonObject("config").getString("consumer_byte_rate"), is("2000"));
 
+        quotas.setConsumerByteRate(3000);
+        quotas.setProducerByteRate(4000);
         original = new JsonObject().put("version", 1).put("config", new JsonObject().put("consumer_byte_rate", "1000").put("producer_byte_rate", "2000"));
-        updated = new JsonObject(new String(kuq.updateUserJson(original.encode().getBytes(Charset.defaultCharset()), new JsonObject().put("consumer_byte_rate", "3000").put("producer_byte_rate", "4000")), Charset.defaultCharset()));
+        updated = new JsonObject(new String(kuq.updateUserJson(original.encode().getBytes(Charset.defaultCharset()), quotas), Charset.defaultCharset()));
         assertThat(updated.getJsonObject("config").getString("consumer_byte_rate"), is(notNullValue()));
         assertThat(updated.getJsonObject("config").getString("producer_byte_rate"), is(notNullValue()));
         assertThat(updated.getJsonObject("config").getString("producer_byte_rate"), is("4000"));
         assertThat(updated.getJsonObject("config").getString("consumer_byte_rate"), is("3000"));
 
         original = new JsonObject().put("version", 1).put("config", new JsonObject());
-        updated = new JsonObject(new String(kuq.updateUserJson(original.encode().getBytes(Charset.defaultCharset()), new JsonObject().put("consumer_byte_rate", "1000")), Charset.defaultCharset()));
+        quotas.setProducerByteRate(null);
+        quotas.setConsumerByteRate(1000);
+        updated = new JsonObject(new String(kuq.updateUserJson(original.encode().getBytes(Charset.defaultCharset()), quotas), Charset.defaultCharset()));
         assertThat(updated.getJsonObject("config").getString("consumer_byte_rate"), is(notNullValue()));
     }
 }
